+++ src/app/src/util/propTypes.js
@@ -13,6 +13,7 @@ import {
     POINT,
     facilityMatchStatusChoicesEnum,
     CLAIM_A_FACILITY,
+    VECTOR_TILE,
     facilityClaimStatusChoicesEnum,
 } from './constants';
 
@@ -212,7 +213,7 @@ export const filtersPropType = shape({
 export const facilityListItemStatusPropType =
     oneOf(Object.values(facilityListItemStatusChoicesEnum).concat('Status'));
 
-export const featureFlagPropType = oneOf([CLAIM_A_FACILITY]);
+export const featureFlagPropType = oneOf([CLAIM_A_FACILITY, VECTOR_TILE]);
 
 export const facilityClaimsListPropType = arrayOf(shape({
     created_at: string.isRequired,