+++ src/django/oar/urls.py
@@ -66,6 +66,9 @@ internal_apis = [
         name='api_token_auth'),
     url(r'^api-feature-flags', views.api_feature_flags,
         name='api_feature_flags'),
+    path(r'tile/<layer>/<cachekey>/<int:z>/<int:x>/<int:y>.<ext>',
+         views.get_tile, name='tile'),
+    url(r'^api/current_tile_cache_key', views.current_tile_cache_key),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
 
 urlpatterns = public_apis + internal_apis